requirements.txt
# TikTok Shop AI Sync Bot - Dependencies

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0

# Shopify API
shopify-python-api>=12.0.0

# Scheduling
APScheduler>=3.10.4

# Testing
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Optional: Enhanced logging and notifications
# slack-sdk>=3.26.0  # For Slack notifications
# sendgrid>=6.11.0    # For email notifications

